Move connexion/ to ir/
