+++ components/PopinManager.jsx
@@ -9,7 +9,7 @@ import { Fragment, PureComponent } from "react";
 
 import { closeCurrentPopin } from "../redux/actions";
 import AjouterCasTypes from "./ir/ajouter-cas-types";
-import LoginForm from "./connexion";
+import LoginForm from "./ir/connexion";
 import EnSavoirPlus from "./ir/en-savoir-plus";
 import { LoginPopin, LogoutPopin } from "./ir";
